+++ srcs/builtins/hash/get_hash.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/02 15:01:50 by arsciand          #+#    #+#             */
-/*   Updated: 2020/02/07 02:02:32 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/06 17:24:35 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ u_int32_t	get_hash(const char *line, u_int32_t size)
 	hash = 0;
 	while (i < ft_strlen(line))
 	{
-		hash = ((hash << 8) + line[i]) % size;
+		hash = ((hash << 8) + (u_int8_t)line[i]) % size;
 		i++;
 	}
 	return (hash);
