+++ srcs/builtins/set.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   set.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mpivet-p <mpivet-p@student.42.fr>          +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/28 22:37:53 by mpivet-p          #+#    #+#             */
-/*   Updated: 2020/02/17 19:05:02 by mpivet-p         ###   ########.fr       */
+/*   Updated: 2020/03/05 04:30:30 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sh42.h"
 
-static int8_t	parse_set(int argc, char **argv)
+static int8_t	parse_set(size_t argc, char **argv)
 {
 	u_int64_t	options;
 
-	options = ft_get_options(argc, argv, "");
+	options = ft_get_options((int)argc, argv, "");
 	if (options & (1ULL << 63))
 	{
 		print_usage("set", options % 128, "set [arg ...]");
@@ -41,9 +41,9 @@ static void		print_internal_vars(t_core *shell)
 	}
 }
 
-static int8_t	new_positional_var(t_core *shell, const char *arg, int count)
+static int8_t	new_positional_var(t_core *shell, const char *arg, size_t count)
 {
-	shell->db.key = ft_itoa(count);
+	shell->db.key = ft_itoa((int32_t)count);
 	shell->db.value = ft_strdup(arg);
 	shell->db.type = SPECIAL_VAR;
 	if (!shell->db.key || !shell->db.value || ft_lstappend(&shell->pos_vars
@@ -56,9 +56,10 @@ static int8_t	new_positional_var(t_core *shell, const char *arg, int count)
 	return (SUCCESS);
 }
 
-static int8_t	get_positional_vars(t_core *shell, t_process *process, int argc)
+static int8_t	get_positional_vars
+	(t_core *shell, t_process *process, size_t argc)
 {
-	int		i;
+	size_t	i;
 
 	i = (process->av[1][0] != '-') ? 1 : 2;
 	free_db(shell->pos_vars);
@@ -77,7 +78,7 @@ static int8_t	get_positional_vars(t_core *shell, t_process *process, int argc)
 int8_t			builtin_set(t_core *shell, t_process *process)
 {
 	int8_t	parsing_ret;
-	int		argc;
+	size_t	argc;
 
 	argc = ft_tablen(process->av);
 	if ((parsing_ret = parse_set(argc, process->av)) > 0)
